usecase 2.8 complete

+++ src/extornal/payment/bankAccount.java
@@ -6,22 +6,22 @@ import java.util.List;
 public class bankAccount {
 
 	public String name ="";
-	protected int sum = 0;
-	protected List<Integer> transactions;
+	protected double sum = 0;
+	protected List<Double> transactions;
 	
 	public bankAccount(String name) {
 		this.name = name;
-		transactions = new LinkedList<Integer>();
+		transactions = new LinkedList<Double>();
 	}
 	
 	public bankAccount(String name,int startwith)
 	{
 		this.name = name;
 		sum = startwith;
-		transactions = new LinkedList<Integer>();
+		transactions = new LinkedList<Double>();
 	}
 	
-	public void getMoney(int amount) {
+	public void getMoney(double amount) {
 		sum += amount;
 		transactions.add(amount);
 	}
